feat: Update UI colors and icons
Introduces a light theme option by changing the default background color to a lighter shade and updates the `backgroundColor` and `darkBackgroundColor` properties in the UI configuration. Also, refactors icon components to use a more streamlined approach and adds new icons. This change aims to improve the visual appeal and flexibility of the application's theme.